FioLab 0.41.0: maturidade e integridade de arquivos
Adiciona validação central de integridade para matrizes e projetos, limites seguros de importação e testes de regressão sem alterar o motor de bordado.

+++ app/src/main/java/com/timachado/fiolab/core/embroidery/EmbroideryModels.kt
@@ -23,8 +23,21 @@ data class EmbroideryBounds(
     val minYUnits: Int,
     val maxYUnits: Int
 ) {
-    val widthMm: Float get() = (maxXUnits - minXUnits) / 10f
-    val heightMm: Float get() = (maxYUnits - minYUnits) / 10f
+    val widthMm: Float
+        get() =
+            (
+                maxXUnits.toLong() -
+                    minXUnits.toLong()
+                ) /
+                10f
+
+    val heightMm: Float
+        get() =
+            (
+                maxYUnits.toLong() -
+                    minYUnits.toLong()
+                ) /
+                10f
 }
 
 data class MachineFinishingInfo(
